Fix TimerClockDivider and remove ClockDivider
